requirements.txt
jsonrpclib-pelix
pika
pymongo
python-dateutil
requests
redis
lxml
scipy
sklearn
newspaper3k
numpy
